Move auth offline-token to a separate package
