+++ cmd/cli/app/auth/auth_login.go
@@ -65,7 +65,7 @@ func LoginCommand(cmd *cobra.Command, _ []string) error {
 
 	// wait for the token to be received
 	var loginErr loginError
-	token, err := login(ctx, cmd, clientConfig, nil, skipBrowser)
+	token, err := Login(ctx, cmd, clientConfig, nil, skipBrowser)
 	if errors.As(err, &loginErr) {
 		if loginErr.isAccessDenied() {
 			cmd.Println("Access denied. Please run the command again and accept the terms and conditions.")
